requirements.txt
flask>=3.0
praat-parselmouth>=0.4
librosa>=0.10
soundfile>=0.12
scipy>=1.11
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
nolds>=0.5
matplotlib>=3.7
werkzeug>=3.0
# v2 training only (optional; core app works without these)
xgboost>=2.0
lightgbm>=4.0
optuna>=3.4
